Add windows support and move compiled files out of src folder
