+++ avs-helper/avs-token-helper.js
@@ -1,27 +1,24 @@
-// tslint:disable-next-line:no-require-imports
-import opn = require("opn");
-import { prompt } from "./common";
-
-const generateQuery = (params: { [key: string]: string }) =>
-    Object.keys(params)
-        .map((key: string) => key + "=" + params[key])
-        .join("&");
-
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const opn = require("opn");
+const common_1 = require("./common");
+const generateQuery = (params) => Object.keys(params)
+    .map((key) => key + "=" + params[key])
+    .join("&");
 console.log("Welcome to the Amazon auth helper!");
 console.log("I will need three things from you:");
 console.log("- Client ID");
 console.log("- Product ID");
 console.log("- Allowed return URL");
-
-prompt("[Press any key to continue]").then(() => {
-    prompt("Client ID?").then(clientId => {
-        prompt("Product Id?").then(productId => {
-            prompt("Redirect URI (allowed return URL)?").then(redirectURI => {
+common_1.prompt("[Press any key to continue]").then(() => {
+    common_1.prompt("Client ID?").then(clientId => {
+        common_1.prompt("Product Id?").then(productId => {
+            common_1.prompt("Redirect URI (allowed return URL)?").then(redirectURI => {
                 const scopeData = {
                     "alexa:all": {
                         productID: productId,
                         productInstanceAttributes: {
-                            deviceSerialNumber: 123, // can be anything
+                            deviceSerialNumber: 123,
                         },
                     },
                 };
@@ -32,19 +29,14 @@ prompt("[Press any key to continue]").then(() => {
                     response_type: "code",
                     redirect_uri: redirectURI,
                 });
-
                 const authUrl = `https://www.amazon.com/ap/oa?${getParams}`;
-
                 if (process.platform === "win32") {
-                    prompt(
-                        "Great! Next I will show you a link. Copy and paste this into your browser to authorize your Alexa product\n[Press any key to continue]",
-                    ).then(() => {
+                    common_1.prompt("Great! Next I will now show you a link. Copy and paste this into your browser to authorize your Alexa product\n[Press any key to continue]").then(() => {
                         console.log(authUrl);
                     });
-                } else {
-                    prompt(
-                        "Great! Next I will now open a browser where you can authorize your Alexa product\n[Press any key to continue]",
-                    ).then(() => {
+                }
+                else {
+                    common_1.prompt("Great! Next I will now open a browser where you can authorize your Alexa product\n[Press any key to continue]").then(() => {
                         opn(authUrl).then(() => {
                             process.exit();
                         });
