+++ package.json
@@ -13,8 +13,9 @@
         "test-rec": "bash test-rec.sh",
         "spell": "cspell src/**/*.ts",
         "style": "prettier -l src/**/*.ts",
-        "avs-token-helper": "node src/avs-token-helper/avs-token-helper.js",
-        "avs-refreshToken-helper": "node src/avs-token-helper/avs-refreshToken-helper.js"
+        "avs-token-helper": "node avs-helper/avs-token-helper.js",
+        "avs-refreshToken-helper": "node avs-helper/avs-refreshToken-helper.js",
+        "build.avs-helper": "tsc --p tsconfig.avs-helper.json"
     },
     "author": "Dolan Miu",
     "license": "MIT",
